Module const_type_layout::inhabited

Helper module to compute whether a combination of types implementing crate::TypeLayout are inhabited or uninhabited.

Macros

all

Helper macro to compute whether all of a list of types, all implementing crate::TypeLayout, e.g. [T, U, V], are inhabited. For instance, a struct is inhabited iff all of its fields are inhabited. The empty list of types is inhabited. This macro resolves into either Inhabited or Uninhabited.

any

Helper macro to compute whether any of a list of types, all implementing crate::TypeLayout, e.g. [T, U, V], is inhabited. For instance, an enum is inhabited iff any of its variants is inhabited. The empty list of types is uninhabited. This macro resolves into either Inhabited or Uninhabited.

Structs

Inhabited

Marker type used to specify that a type implementing crate::TypeLayout::Inhabited is inhabited.

Enums

Uninhabited

Marker type used to specify that a type implementing crate::TypeLayout::Inhabited is uninhabited.