const_type_layout
Module typeset

Helper module to compute the set of types that a type links to and expand it into the complete type graph.

Macros

tset

Helper macro to expand a list of types, e.g. H, R1, R2, and an optional tail, .. @ T, into a set of types.

Traits

ComputeTypeSet

Computes the set of types that a type links to.