const_type_layout::typeset
Macro tset

pub macro tset {
    () => { ... },
    (.. @ $T:tt) => { ... },
    ($H:ty $(, $R:ty)*) => { ... },
    ($H:ty, $($R:ty,)* .. @ $T:ty ) => { ... },
}

Helper macro to expand a list of types, e.g. H, R1, R2, and an optional tail, .. @ T, into a set of types.

This macro is used when implementing the ComputeTypeSet::Output associated type to specify the list of types a type links to.