Trait const_type_layout::typeset::ComputeTypeSet

pub unsafe trait ComputeTypeSet: TypeLayout {
    type Output<T: ExpandTypeSet>: ExpandTypeSet;
}

Available on crate feature derive only.

Computes the set of types that a type links to.

Safety

It is only safe to implement this trait if it accurately includes all inner component types that are referenced by this type’s layout. Use #[derive(TypeLayout)] instead.

Example

The struct Foo with u8 and u16 fields links to u8 and u16:

struct Foo {
    a: u8,
    b: u16,
}

unsafe impl ComputeTypeSet for Foo {
    type Output<T: ExpandTypeSet> = tset![u8, u16];
}

Note that to you implement ComputeTypeSet you must also implement crate::TypeLayout for it.

Required Associated Types

type Output<T: ExpandTypeSet>: ExpandTypeSet

Extend the set T into a (larger) set containing also the types this type links to.

Enums implementing crate::TypeLayout and ComputeTypeSet manually should include crate::ExtractDiscriminant::Discriminant in their ComputeTypeSet::Output using the tset helper macro.

Object Safety

This trait is not object safe.