const_type_layout
Trait TypeLayout

pub unsafe trait TypeLayout: Sized {
    const INHABITED: MaybeUninhabited;
    const TYPE_LAYOUT: TypeLayoutInfo<'static>;
}

Utility trait that provides the shallow layout of a type.

Safety

It is only safe to implement this trait if it accurately describes the type’s layout. Use #[derive(TypeLayout)] instead.

Example

The struct Foo with u8 and u16 fields implements TypeLayout as follows:

struct Foo {
    a: u8,
    b: u16,
}

unsafe impl TypeLayout for Foo {
    const INHABITED: MaybeUninhabited = inhabited::all![u8, u16];

    const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
        name: ::core::any::type_name::<Self>(),
        size: ::core::mem::size_of::<Self>(),
        alignment: ::core::mem::align_of::<Self>(),
        structure: TypeStructure::Struct {
            repr: "",
            fields: &[
                Field {
                    name: "a",
                    offset: MaybeUninhabited::new::<u8>(::core::mem::offset_of!(Self, a)),
                    ty: ::core::any::type_name::<u8>(),
                },
                Field {
                    name: "b",
                    offset: MaybeUninhabited::new::<u16>(::core::mem::offset_of!(Self, b)),
                    ty: ::core::any::type_name::<u16>(),
                },
            ],
        },
    };
}

Note that if you implement TypeLayout, you should also implement typeset::ComputeTypeSet for it.

Required Associated Constants

const INHABITED: MaybeUninhabited

Marker for whether the type is inhabited or uninhabited.

const TYPE_LAYOUT: TypeLayoutInfo<'static>

Shallow layout of the type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.