Trait const_type_layout::TypeLayout

pub unsafe trait TypeLayout: Sized {
    type Inhabited: OutputMaybeInhabited;

    const TYPE_LAYOUT: TypeLayoutInfo<'static>;
}

Available on crate feature derive only.

Utility trait that provides the shallow layout of a type.

Safety

It is only safe to implement this trait if it accurately describes the type’s layout. Use #[derive(TypeLayout)] instead.

Example

The struct Foo with u8 and u16 fields implements TypeLayout as follows:

struct Foo {
    a: u8,
    b: u16,
}

unsafe impl TypeLayout for Foo {
    type Inhabited = inhabited::all![u8, u16];

    const TYPE_LAYOUT: TypeLayoutInfo<'static> = TypeLayoutInfo {
        name: ::core::any::type_name::<Self>(),
        size: ::core::mem::size_of::<Self>(),
        alignment: ::core::mem::align_of::<Self>(),
        structure: TypeStructure::Struct {
            repr: "",
            fields: &[
                Field {
                    name: "a",
                    offset: MaybeUninhabited::new::<u8>(::core::mem::offset_of!(Self, a)),
                    ty: ::core::any::type_name::<u8>(),
                },
                Field {
                    name: "b",
                    offset: MaybeUninhabited::new::<u16>(::core::mem::offset_of!(Self, b)),
                    ty: ::core::any::type_name::<u16>(),
                },
            ],
        },
    };
}

Note that if you implement TypeLayout, you should also implement typeset::ComputeTypeSet for it.

Required Associated Types

type Inhabited: OutputMaybeInhabited

Marker type for whether the type is inhabited or uninhabited. The associated type must be either inhabited::Inhabited or inhabited::Uninhabited.

Required Associated Constants

const TYPE_LAYOUT: TypeLayoutInfo<'static>

Shallow layout of the type.

Object Safety

This trait is not object safe.